Add first version of UI library
This change adds the first version of the Generic UI library/framework
(GUI). The library strives to be completely independent of the
underlying architecture. That is, it is compatible with windowing as
well as terminal based systems and it does not care about the specifics
of any GUI toolkit used.
Part of this story means that it is not an out-of-the-box replacement
for your favorite user interface toolkit, but that it should be seen
more as a framework, providing certain hooks and basic functionality
typically seen in a UI, such as dispatching of events. To make proper
use of it, the functionality provided by this library needs to be glued
with the underlying system.