Add the notion of a root widget
So far the UI supported the addition of widgets to it. However, these
stand in no relation to each other. Conceptually, a user interface
comprised of widgets typically forms some sort of hierarchical
structure, with one entity at the root.
Although it would make sense to place our Ui struct itself at the root,
architecture-wise that is not a feasible approach. Instead, we want a
certain widget to be special in the sense that it forms the root of the
hierarchy. It is special because it is the only widget that does not
have a parent widget.
To that end, this change adds support for adding a root widget to the
UI. The constructor of such a widget cannot accept an ID of the parent
widget.