Add support for rendering the UI
This patch adds support for rendering the UI. To accomplish this task
it makes two main changes: First, it introduces a new trait, Renderable,
which represents objects that can be rendered. Second, in addition to
the existing Widget trait, which is concerned with ID management and
parent/child handling, it introduces UiWidget, a composition of traits
that need to be implemented for objects that are to be managed by the
UI.
This multi-trait design may seem over engineered as of now, but it will
be the basis of the overall architecture and be a requirement for
subsequent changes.