Introduce a dedicated Renderer
So far rendering of widgets was possible, but very inflexible. In
particular, a widget's rendering code would be implemented inside the
widget itself, leading to a very fixed representation of the object.
We need a more flexible approach where the rendering code for a certain
type of widget can be swapped out dynamically, at build time or at run
time. We also do not want a widget's rendering code to be bound to the
widget, as that would make it usable only in conjunction with the
underlying rendering system it was written for. Rather, widgets should
be abstract entities that specify a behavior, but whose representation
is dependent on the drawing toolkit used.
To that end, this change introduces the `Renderer` trait. Object of this
trait are capable of rendering widgets, while different renderers may
display the same widget differently. Renderers naturally need to know
which widget they can render and in the currently envisioned design the
widgets get passed in as `Any` and it is up to the renderer to figure
out the actual type and render it accordingly. It is also up to the
renderer to panic on unknown widgets or simply draw them in a generic
way.