Add support for focusing widgets
Right now one of the most important use cases of the UI evolves around
the dispatching of input to the correct widget. As a first step, only
key related input will matter and this is typically dispatched to the
widget currently being focused.
To support such a work flow, this patch adds the notion of a focused
widget to the UI. In our case, that is achieved by storing the ID of the
widget currently having the focus in the Ui struct.