Add basic event handling support
This change adds basic support for handling events to the library. In
particular, it introduces a new trait, Handleable, that is to be
implemented by each widget registered with the UI. The only function in
this trait, handle, accepts a single event to process. At this point we
only support key related events (key presses and releases; and only in a
limited fashion, i.e., without modifier keys etc.). These events are
forwarded to the currently focused widget.