Add support for focus events
One of the most important features of the UI is to handle focusing of
widgets properly and to allow for changing the focus, such that
different widgets can receive inputs from the keyboard.
To facilitate this use case, this change adds support for focus events.
A widget can emit such an event in order to focus another widget (the
event can also come from the outside, but that is irrelevant). Because
focusing is handled by the UI itself, this operation is represented as
an event and cannot be a simple function call (widgets do not have
access to the UI directly).
Furthermore, focus events are only meaningful to the UI itself. As such,
part of this change is the introduction of a new event class, UiEvent.
UiEvent objects form a true superset of the already present Event
objects. Widgets can handle normal an ordinary Event, but not a UiEvent.
However, they can emit a UiEvent.