Introduce gui-derive crate
Implementing a full-fledged UI widget is quickly becoming a bloated
process involving a lot of boilerplate code. That is mostly due to the
various traits that have to be implemented for such a widget.
Luckily, there is a mechanism for coping with exactly this problem:
the custom derive functionality allows for the implementation of certain
traits in a concise way, by means of the `#[derive(x)]` syntax that
annotates a struct (or enum).
This change introduces a new crate, gui-derive, that provides a custom
derive macro that allows for this sort of simplification. The
requirement for a separate crate stems from the Rust language support
side of things.
Note that with this patch we only provide the first building block for
such a macro: support for parsing macro attributes, which will be used
for identifying which higher level widget to derive from. The remaining
functionality will follow.