Introduce GuiWidget custom derive functionality
With this patch we introduce a procedural macro providing automated
derive support for Widget -- the composition of the Object, Renderable,
and Handleable traits.
Note that only the Widget and Renderable traits will be implemented
automatically. Handleable needs to be provided by the user as event
handling logic is specific to each and every widget.