Check for required struct fields
The error messages reported for procedural macros are not the best and
debugging problems is cumbersome. We cannot do too much about that, but
in an attempt to improve things at least a bit, this change adds
automated checks for the required struct fields to the GuiWidget custom
derive macro.
If a required field is missing a panic will occur (as is usual for
procedural macros) that signals the field name as well as the required
type. Note that we do not check for the correct type ourselves as that
is hardly doable correctly on a token stream without context.