Add support for creating a default new() implementation
If a widget does not contain custom fields in addition to the ones
required by the respective widget type (e.g., id and parent_id), it is
possible to create a default implementation of the new() function for
the derived type.
This change adjusts the macro expansion logic to be able to generate
such a default implementation. This functionality has to be explicitly
solicited by means of the newly introduced #[GuiDefaultNew] attribute.