Add custom derive macro for gui::Handleable
Sometimes, especially for testing purposes and quick prototyping, a
widget does not require a custom event handler. In those cases the need
for implementing the gui::Handleable trait becomes a burden.
To alleviate this problem, this change introduces another custom derive
macro that automatically creates an empty (and, thus, pretty much
useless for most intents and purposes) implementation of the trait.