Do not error out for unknown attributes
As it turns out all attributes for a struct are passed to the
implementation of the custom derive procedural macro. Right now we
cannot handle such cases as we error out for any unknown (for the macro)
attribute, e.g., #[allow(dead_code)].
This change adjusts the logic to no longer cause an error when
encountering an unknown attribute, but rather ignore it silently.