Change #[GuiDefaultNew] attribute to #[gui(default_new)]
The #[GuiDefaultNew] style attributes are (perhaps subjectively) not a
very nice solution. They clutter the global attribute namespace
unnecessarily and each subsequently added item will add to that. It also
is not very consistent with other attributes (serde, for exmaple, uses
attributes of the form #[serde(...)].
As a proposed improvement, this patch renames the #[GuiDefaultNew]
attribute to #[gui(default_new)]. This syntax is seemingly more
prevalent and allows us to have the all custom attributes behind a
common name.
Note that we do not touch the #[GuiType = "xxx"] attributes in this
change. There exists a different plan for those.