Introduce dedicated macros for different widget types
There exists another problem in our macro infrastructure and that is
with the #[GuiType = "xxx"] attribute. In retrospect, having a dedicated
attribute to specify the "type" of widget seems sub par. That reason
along with the problems of attribute namespace pollution as well as
uniformity with how other crates behave caused a rescan of the design
space.
This change replaces the GuiType attribute with explicit derive macros
for the individual types. This appears to be the cleaner solution. In
addition, this refactoring actually yields a good amount of code
simplification, rendering a couple of tests superfluous and allowing for
their removal.