Add support for derive from generic structs
Our current set of derive primitives has one problem: it does not work
for structs that contain some form of generic parameters. In that case
wrong sequences of tokens would be generated, e.g., `struct <T> <R>
Foo`.
But the scary thing is, this does not seem to be a problem specific to
this library and yet there does not seem to be anybody addressing it (or
perhaps they just hide very well).
Anyway, the solution to this problem is to mess with the syntax tree in
order to produce a valid list of tokens. This change accomplishes this
task. Both the implementations of the gui::Renderer and gui::Widget
traits depend on proper handling of generics and they are adjusted
accordingly.