Introduce Quit event
Sometimes it is necessary for widgets within the UI to communicate
something to the outside. An essential example is the termination of the
program by press of a button.
We currently lack a generalized communication channel of that sort and
it is questionable whether establishing one is feasible or desired.
However, termination of the program is a very basic need that can arise
in every sort of UI. For that reason this change adds a new event type,
Quit, to the UI that is bubbled up to callers who are subsequently able
to react on it.