Introduce pre- and post-render steps
In many scenarios it is a requirement for the renderer to perform some
pre-render setup (e.g., setting of texture) and/or some post-render tear
down (e.g., swapping display buffers).
With this change we add pre_render() and post_render() methods to the
Renderer trait, enabling such use cases. The functionality is triggered
by the Ui object on its render invocation before and after the rendering
of the individual widgets, respectively, and as such is fairly
hardwired. However, each individual renderer object can fill in the
individual methods according to its requirements.