Introduce generic UI event handler function
In anticipation of more UI events being added, this change marks a small
refactoring to make the code easier to extend in the future. In
particular, it factors out a new function, handle_ui_specific_event,
that is used to share code on in two locations in the ui.rs file.