Make test widgets easier configurable
In order to test event handling we introduced a special type of widget,
TestEventWidget, that contained functionality tailored to testing
events. When we want to test more event related behaviors in the future,
this approach simply does not pan out: we cannot introduce a new type
for each configuration we want to test. We already have this problem
with the different widget types (root widget, container, widget), only
one of which is special cased like that currently.
This change rethinks this functionality by reworking the common test
widgets we have to accept a handler function that can be used to
customize a widget's behavior at the object level. By doing so, it
renders TestEventWidget redundant.
Note that the implementation is fairly cumbersome due to the requirement
of having stateful handler functions: we cannot use compile time defined
functions but need to rely on closures. These in turn do not have a
fixed size and so we have to box them up. Because boxes do not implement
the Debug trait (a requirement for the gui crate), we have to define a
custom struct wrapping it. Sigh.