Introduce new Custom event type
We cannot predict all possible events that clients would like to
support, and so some form of extensible infrastructure needs to be
provided that caters to at least the majority of conceivable use cases.
It is possible, for example, that one widget needs to inform another
about an update.
Furthermore, currently there exists no reasonable way to get data back
out of a widget. Once it is in there it is there to stay, and sharing
through a reference a not possible as it stands.
To solve both issues, this change piggy backs on the existing event
infrastructure and introduces a custom event which can carry arbitrary
data. Using such an event we open the gates for implementing more domain
specific events but also provide a way to retrieve data back from a
widget.
On the implementation side we rely on a boxed std::any::Any object
within the custom event to contain the payload data. Unfortunately,
embedding an Any object in the event enums has the side effect of no
longer allowing for default clone and comparison functionality to be
derived. Right now that is only a problem for the tests, and we do not
expect clients to rely on such functionality, but we have to jump
through quite a few hoops to get the tests back. In particular, most of
the event related tests are integration tests which, by design, are
built against the crate in the normal configuration. We cannot provide
custom implementations of those traits specifically for tests (i.e.,
guarded by #[cfg(test)] or similar means). So we end up providing
functions that offer the required deep clone and compare functionality
without being able to rely on the nice syntactic sugar present when
providing this functionality through a trait impl.