Remove generic type parameter from add_*widget
An upcoming change will introduce a trait that is to be implemented by
the Ui struct, the goal of which being the ability to create trait
objects from objects of said struct. However, in order to enable this
functionality we need to make the struct object safe first, which means
eliminating the generics from the signature of the add_widget and
add_root_widget methods.
This change pursues this endeavor by introducing types representing the
functions we accept in both contexts.