Add support for creating widgets recursively
Although we have always supported dynamic creation of widgets, doing
so has also been fairly constrained: because a Ui object is required to
create a new widget, this functionality was only accessible from outside
of a widget itself, as widgets do not know about their containing Ui.
That means that an arguably natural way of thinking about the formation
of a user interface was not possible: widgets creating and managing
other widgets from within themselves.

This change alleviates this restriction by allowing a widget to create a
new widget as part of its construction. We enable this use case through
a "capability" supplied to the constructor: an object specifically
exposing various widget related operations, including add_widget.

On the implementation side this capability is nothing more than a trait
that is implemented by the Ui struct. Hence, a pointer to such a trait
object suffices for the creation of widgets while also preventing other
Ui functionality such as the render() or handle() methods from being
invoked.
This design was found to be much nicer in terms of code sharing and
conciseness than having a dedicated struct that is created on demand --
a design which was used initially.