Use dynamic dispatch for Renderer in Renderable.render
The Renderable trait, as originally envisioned, used static dispatch
(i.e., generics) to invoke a method on the passed in Renderer. While
generally preferable, doing so was just found to create a huge mess.
Most importantly, that is because we want the trait to be object safe,
we cannot use generics on the render method itself, but have to push
them up to the struct level. This pollutes all users of this trait with
generics for this renderer.
With this change we use a dynamic approach in which we just pass in a
trait object representing the Renderer to use. Doing so has the
additional advantage that we are now at least conceptually able to
render a UI using multiple renderers.