Declare widget creation functions as FnMut
The move away from using a generic for the widget creation function in
the Ui.add_widget method has been shown to be problematic: Because we
use an Fn we capture the environment in a read-only manner and that may
be insufficient for the construction of certain widgets. Unfortunately,
we cannot go back to using FnOnce because that would require knowing the
size which in turn would only be possible in conjunction with generics
which we cannot use because of the trait object requirement that
necessitated the change in the first place. Alternatively, boxing up the
FnOnce would work conceptually, but the language design does not
currently support this case properly either (there have been long
discussions around the introduction of FnBox).
With this change we alleviate the problem mostly. Specifically, we
declare the widget creation functions as &mut FnMut, meaning we have
mutable access to the environment. This still does not provide
moveability directly, but it can be emulated on top using an Option as a
parameter and panicking if the supplied value is None.