Require root widget creation function in Ui constructor
By design, there can only be one root widget in a UI. Similarly, it can
be argued that a UI without a root widget is useless as nothing can be
done with it.
For these reasons, this change merges the step of instantiating and
registering the root widget into the Ui object's constructor. By doing
so we also eliminate the case of repeated invocation of add_root_widget
at the type level.