Remove usage of Cell and RefCell from Ui struct
With the most recent refactoring it became apparent that we do not
really need the runtime based borrowing provided by RefCell and Cell.
Hence, with this change we remove the usage of those wrappers around the
boxed widgets and the focus attribute.
This step simplifies the code once more and get it more in line with
what idiomatic Rust should look like.