Introduce WidgetRef trait
With an anticipated future change we want give clients the possibility
of supplying widget IDs or true widget references to the Ui (and,
potentially more importantly, the Cap in general).
The background is that sometimes an ID is all that is present while
in other cases it will be necessary to supply a true reference, namely
then when a call originates in a widget and targets the very same
widget.
To abstract away from the respective actual type, this change introduces
the WidgetRef trait. This trait is implemented by individual Widgets as
well as the Id struct and allows for conversion between the them.