Implement std::fmt::Display trait for Id struct
With an upcoming change we want to format an Id nicely when panicking.
Since those formats are happening internally to the Ui we could
reference the idx member directly. However, we may as well implement the
std::fmt::Display trait instead, which would benefit clients who want
to print an Id as well.