Make widgets vector contain Options of boxed Widgets
So far we have gotten away with borrowing one or more widgets
non-mutably or acquiring a single mutable borrow of a single widget and
with it the Ui.
However, with an upcoming change this will no longer be enough: we will
require multiple mutable borrows of the Ui object because we want to
enable widgets to mutate other widgets (within *very* strict
boundaries). Of course, this is not possible directly with safe Rust.
The escape hatch we implement with this change is the storage of widgets
within an Option. This step allows us to replace the widget temporarily
with a None value while effectively taking ownership of the widget (and
with that gaining mutable access to it). Hence, we are able to access
the Ui mutably while also having a mutable widget.