Provide a Cap object to Handleable::handle
With a recent change we provided a Cap object to a widget's constructor,
effectively enabling recursive creation of widgets -- arguably a
necessary property of a widget based user interface. However, there is
another case where we may want to dynamically create a widget: when an
event arrives.
This patch enables this use case. We use the same mechanism we employed
before: we pass a Cap object to a Handleable's handle method. However, a
problem arises due to the fact that the we have a mutable widget borrow
(effectively resulting in a mutable Ui borrow) while also requiring a
mutable Cap (yet another mutable Ui reference). To cope with this
problem we temporarily take ownership of the widget by taking it out of
the Option we store it in in the first place. Doing so means that now we
would panic if we were to access the widget recursively (because that's
what we do if the Option is None). The only way this can happen is if
the widget's handle method uses the provided Cap object. We fudge that
case by requiring that the widget supply its own reference to the Cap
object, and not its own Id. This way we do not have to lookup the widget
and, hence, do not panic. This use case is enabled since all methods in
the Cap trait accept a WidgetRef, which can be either an Id or an actual
reference.