Add test for using Cap from handle method
This change adds two tests for using the Cap object that is now passed
to a Handleable's handle method. The first test verifies that we can
dynamically create a widget from the event handler. The second test
checks that we panic as expected when using an Id as the WidgetRef that
is the calling widget's Id.