Update syn crate to 0.12
This change updates the syn crate to version 0.12 as well as the quote
crate to 0.4. As part of this version bump we have a new dependency to
proc-macro2.
This is an intermediate step towards updating syn to the most recent
version. The crate has experienced major refactorings of its public
interface, necessitating a step-wise update.