Add support for more keys
When we initially added support for events and with that key press and
key release events, We started off with a very minimal set of keys being
supported. To make the library more usable we need a bunch of more keys
that are typically used by program.
And so while not being exhaustive (functions keys and modifiers are
still missing, for example), this change adds support for a good amount
of commonly used keys to the crate.
It is also noteworthy that with this change the "return" key is
represented as a dedicated enum value as opposed to a simply the '\n'
character. The reason for this change is that we want to make sure that
return is a first class citizen with proper and unambiguous
representation. Being a character, it is prone to being represented
differently (e.g., as '\r') or also just being written out directly,
which may or may not be the proper way to end a line, depending on the
platform the program runs on. By representing this key as a dedicated
value we hopefully make users think about those cases more.