Detect wrong sharing of Id objects between different Ui objects early
One of the constraints we imposed on ourselves was to have as little
global state as possible in the UI. Among other things, this allows us
to create multiple Ui instances that are conceptually isolated from each
other. One can envision that property being used in order to have one Ui
object per display, for example.
One problem we have, though, is the sharing of Id objects between
different Ui objects. We have not found a proper way to catch such
invalid sharing at compile time through the type system and so this
change puts in place additional run time checking infrastructure that
catches misuse.
To that end, we introduce a validation step for all Id objects that
verifies belonging to the correct Ui. This step is only effective on
debug builds and should be fully compiled out otherwise. We also
facilitate the type system to make sure that we validate any user
supplied Id before using it. Note that by design we keep all this
validation logic internal to the Ui. We do not want clients to have to
deal with different type of IDs. That means that right now we perform
more checks than necessary. However, these checks can be circumvented
from within the Ui module should the need arise (i.e., due to
performance reasons), by directly referencing the internal 'idx' member.