Add functionality for chaining events
Sometimes it is a requirement for a widget to react to a given event in
various ways. Consider the case where a widget receives a key event and
subsequently needs to instruct another widget to change its state while
also having to indicate to someone that it itself got updated. Both
interactions may have to piggy back on the event handling infrastructure
but only one event can be send at any given point in time. We could
implicitly make one message to another widget make the other widget emit
another event (assuming it would not already have done so in the first
place, in which we are back at square one) but this is a completely
non-obvious scheme.
Rather, what we really want to express is the ability to emit an
arbitrary number of events from a widget's event handler. We have
explored the design space quite thoroughly, and the following
possibilities were found (in addition to the one briefly mentioned
above):
1) Adjust the signature of all event handling functions and return a
   vector of UiEvent objects from them.
   The downside with this approach is that we now introduce vector code
   to clients even where this is not strictly necessary. We also have no
   way to enforce that a vector contains at least one element and so
   clients can return empty vectors. That is not wrong, but it is also
   not nice. Overall, despite being subjective, returning a vector just
   seems ugly API design.
2) Add another branch in UiEvent that contains a vector of UiEvent
   objects.
   That is nicer than directly returning a vector from all event
   handlers. The problem we have with this approach is that clients can
   create arbitrarily nested events now, that is, they don't really have
   to form a chain in the strict sense. You can have a vector of
   UiEvents where any arbitrary event can in turn comprise a vector of
   UiEvent objects.
3) Rework the event handling entirely and use the capability object for
   sending events in a synchronous manner.
   At first sight this approach seems feasible and, indeed, beautiful,
   as we can directly send events and don't have to go through the
   indirection of returning something first (with all the control flow
   complexities that entails). On second thought that approach cannot
   work because we will eventually run into borrowing conflicts.
   Consider widget 'a' sending something to widget 'b'. That's fine. But
   the moment widget 'b' now sends something back in response it cannot
   supply a true reference to 'a' and so it uses an Id. That causes a
   look up and with that a borrowing violation.
   So an implementation variation would be to just queue up all events
   to be sent in the Cap object. But this scheme gets incredibly complex
   and intricate because now we need to have per-widget book keeping of
   queued up events. That just seems undesirable.
4) Use coroutines and the 'yield' syntax to effectively return multiple
   events from a single event handler, one after the other.
   This would be our ideal solution and would represent the intent very
   naturally. Unfortunately, coroutines are not yet a stable feature
   and, hence, off the table.

In the end we decided in favor of yet another approach: Introduce
another event, MetaEvent, wrapping a UiEvent that essentially can form a
linked list, and with that an event chain. With that design we can
enforce -- at a type level -- that clients cannot run havoc when it
comes to event composition. That is, we can enforce a true chain being
formed through compile time measures.
The downside is that we introduce yet another event type. Naming is
already complicated and this new type just exaggerates the issue. Plus,
we have to adjust all existing event handler signatures.
Nevertheless, this is the approach with the seemingly best trade-offs
and hence, the preferred way.
Note that the chaining functionality provided by MetaEvent object is
meant for UI internal communication only. As such, the Ui object's
handle method (which interfaces with the outside) does not return a
MetaEvent but keeps using UiEvent objects. Hence, each event handled is
eventually transformed into a single resulting event (or none).