Simplify Ui struct's event handling internals
On a second look the Ui object's event handling internals are not very
explicit and to the point. The 'handle' method dispatched into various
sub methods whose behavior and call stack was not very obvious. What we
really want to do in there is to distinguish between events that are to
be delivered to the focused widget and those that are not.
With this change we remove the utility methods that we introduced
earlier and refactor the 'handle' method accordingly.