Introduce concept of a bounding box for rendering
Currently there are no restrictions on where on the screen a widget can
draw. Similarly, there is also no guidance as to where it should draw.
In essence, all this logic resides inside the renderer. Ideally, we
would want each widget to get passed in the area that it can occupy.
This change marks a first step into this direction. It introduces a new
object, a bounding box (BBox for short) that represents a rectangular
area that can be drawn to.
Note that currently there is no enforcement that rendering happens into
this area and generally it will be up to the respective renderer object
to enforce that, but including this logic already has the potential to
simplify code slightly.
Note furthermore that for now the bounding box is representing the
entire screen, irrespective of which widget we render. In the future
each widget will shrink the bounding box representing the area that its
children can occupy.
Lastly, note that this entire mechanism is contained to only the
rendering logic. We want to avoid having to associate a position and
size with every widget for as long as possible (this is likely going to
break once we want/need mouse support).