Allow for proper adjustments of bounding boxes
The previously introduced bounding box was a very static one so far: it
unconditionally represented the size of the medium being drawn to (e.g.,
the screen). This was never meant to be the end of the story, however:
widgets are supposed to adjust the bounding box that is to be used when
rendering their children.
This change adjusts the logic to allow for precisely that. It adjusts
the 'render' method to not only accept but also return such a bounding
box object, which will subsequently be passed to the widget's children,
if any.