Add functionality to remember and retrieve the last focused widget
Sometimes it is necessary to temporarily focus a widget and have that
widget -- after doing its job -- change the focus back to the previously
focused widget. Currently implementing such trivial functionality is not
trivial at all: one would have to redirect all focusing and refocusing
through the root widget. That in turn is only reasonably possible
through the usage of events with the root widget remembering the
previously focused widget. That is a clumsy and error-prone solution to
a trivial problem and as such unacceptable.
To provide a solution in a more straight forward manner, this change
adjusts the Ui itself to provide such functionality. This API should be
regarded as preliminary. The hope is that eventually a more generic
solution covering this and related problems can be found, but proper
design requires some more thought and experimentation. Along with the
retrieval of the last focused widget another method for retrieving the
currently focused widget is also added.