Store basic widget related attributes in the Ui object
One major source of incompatibilities moving forward with the
development of the crate is the addition of methods to the Widget trait
that are backed by attributes inside the user defined widget type
itself. E.g., consider the case that two news methods for indicating and
setting the visibility of a widget are to be introduced to the Widget
trait. These methods are backed by a flag that is a simple boolean value
that would be stored on a per-widget basis alongside the likes of the
widget's ID, its parent's ID, and its children.
While the creation of definitions for those methods would be handled by
the gui-derive crate, the backing attributes pose a bigger problem: each
of them would require an adjustment in client code. The reason this
adjustment is required is because custom derive macros do not allow for
creation or modification of the corresponding struct's attributes. Now,
procedural macros in general should be able to produce those attribute
definitions in a generic manner, but they are still an unstable as of
Rust 1.27 -- making them unusable in the context of this crate.
The alternative use of macro rules to create those definitions, while
conceptually feasible, has been shown to be a significant effort to get
correct for an arbitrary struct in general, as visibility attributes,
life times, and more all have to be parsed and handled correctly.

Pondering over alternative ways out, a much neater solution came to
mind: instead of pushing the burden of definition of those attributes to
the user, these can actually be kept inside the Ui object itself and be
looked up through an Id. That would mean that all client have to put
into their custom widget definition is an Id, which is much easier to
guarantee compatibility for.
To that end, this change marks a first step towards moving the
'parent_id' and 'children' attributes under the management of the Ui
struct. In particular, it introduces a new type, WidgetData, that
contains said attributes. Note that right now these attributes, although
set, are not actively used by widgets. These adjustments will follow in
the form of subsequent changes.