Remove parent_id attribute from all widgets
A previous change introduced the means to store the relevant core widget
attributes such as the ID of the parent widget and the vector of IDs of
its children outside of the struct implementing the Widget trait.
However, but kept those attributes in place.
This change marks a first step towards fully transitioning to using
those newly introduced attributes by removing the parent_id variable
from all widgets. By doing so the constructors of all widgets get
simplified in that they no longer have to accept the ID of the parent as
a direct parameter. Instead, if this ID is needed it can be retrieved
using the Cap object that is available.