Remove children attribute from all widgets
As a next step in the process of moving core widget related attributes
in the Ui object this change also removes the need for the children
field from all widgets.
This change greatly simplifies the custom derive logic we have in place
in that effectively all widgets now behave the same. That is true to the
point that we can actually unify the three custom derive macros into a
single one, but this task will be performed as a follow up change. One
implication of this move is that now all widgets behave as containers
(although it is possible to overwrite the definition of the add_child
method, of course, but that is not the intention). It remains to be seen
whether this behavior has undesired consequences.
With this change we conclude the transition of the attributes
themselves. The only remaining attribute stored in a widget is the
widget's own ID. It is an open question whether we can and want move its
management into the Ui object as well.