Get rid of WidgetRef trait
As it turns out with all this shuffling around of code and moving of
core widget attributes under the umbrella of the Ui object we no longer
require the WidgetRef trait. This trait was initially devised to solve
the case where user code uses a Cap object to inquire additional
information about a widget (directly or indirectly) from a handle method
(which requires a mutable borrow of the widget itself). In order to
retrieve the requested data the Ui would need to borrow the widget,
which panic. By supplying a true reference to the widget in question we
circumvent this problem elegantly.
However, even more elegantly, with the move of relevant attributes into
the Ui object we no longer run into these problems in the first place.
Hence, the WidgetRef trait is no longer necessary. To that end, this
change removes the trait from the two crates, greatly simplifying the
logic once more.