Unify widget derive macros into GuiWidget
With recent changes we no longer make a distinction between the root
widget, container widgets, and "normal" widgets, at least not in terms
of data structures (the root widget still has no parent but that is
solely expressed as a constraint and not by virtue of having different
attributes). What we omitted doing as part of the series of patches
removing this distinction is to actually adjusting the custom derive
logic accordingly.
This change makes good for that neglect by unifying the GuiRootWidget,
GuiContainer, and GuiWidget macros.