Introduce Any-style casting functionality for Widget trait
Pretty much since the inception of the rendering functionality we have
had to cast Widget trait objects into concrete widget types in order to
allow for customized rendering. At the time we were passing around those
widgets in the form of Any objects. That was not very nice, as Any can
literally represent pretty much any type and, hence, failed to convey
the intention of the object being passed in being a widget. Further,
even if the concrete type was of no interest and a Widget enough it was
still necessary to cast.
With this change we change this fact by implementing similar casting
functionality to that of Any directly for the Widget trait. That means
that the rendering code can now finally accept actual widget trait
objects, making it much more expressive.
The reason for this change is that with an upcoming patch we will need
this sort of casting in a second location, and we would have similar
problems as mentioned before for the rendering code. On the downside, we
are no longer a safe crate in the strict sense, as the type casting
functionality is contained in unsafe blocks. The functionality is the
same as that of Any (to the point that it could be replaced one-to-one
with it), so from that point it is safe. However, there is the
possibility of client code overwriting the type ID functionality in an
unsafe manner and, hence, causing havoc.